+++ src/main/java/com/example/myapp/repository/DoctorSpecialtyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DoctorSpecialtyRepository extends JpaRepository<DoctorSpecialty, Integer> {
+public interface DoctorSpecialtyRepository extends JpaRepository<DoctorSpecialty, Long> {
 }
